requirements.txt
Django==4.2
django-extensions==3.2.1
django-crontab==0.7.1
django-cors-headers==4.3.1
python-magic==0.4.27
channels==4.0.0
requests>=2.31.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
cryptography==42.0.5
gunicorn==21.2.0